refactor: relocate EmailAuthMsg implementations
